work

+++ src/main/java/com/robertx22/Stats/StatMod.java
@@ -8,35 +8,35 @@ import com.robertx22.uncommon.utilityclasses.IWeighted;
 
 public abstract class StatMod implements IWeighted, IGUID {
 
-	@Override
-	public int Weight() {
-		return IWeighted.UncommonWeight;
-	}
+    @Override
+    public int Weight() {
+	return IWeighted.UncommonWeight;
+    }
 
-	private static Random ran = new Random();
+    private static Random ran = new Random();
 
-	public abstract Stat GetBaseStat();
+    public abstract Stat GetBaseStat();
 
-	public abstract int Min();
+    public abstract float Min();
 
-	public abstract int Max();
+    public abstract float Max();
 
-	public abstract StatTypes Type();
+    public abstract StatTypes Type();
 
-	public abstract String GUID();
+    public abstract String GUID();
 
-	public int GetValueByPercent(int percent) {
+    public int GetValueByPercent(int percent) {
 
-		return Min() + (Max() - Min()) * percent / 100;
+	return (int) (Min() + (Max() - Min()) * percent / 100);
 
-	}
+    }
 
-	public int GetRandomNumber() {
+    public int GetRandomNumber() {
 
-		int max = Math.round(Max() - Min() + 1);
+	int max = Math.round(Max() - Min() + 1);
 
-		return Min() + ran.nextInt(max);
+	return (int) (Min() + ran.nextInt(max));
 
-	}
+    }
 
 }
