+++ src/main/java/com/robertx22/database/stat_mods/flat/resources/ManaFlat.java
@@ -16,12 +16,12 @@ public class ManaFlat extends StatMod {
 	}
 
 	@Override
-	public int Min() {
+	public float Min() {
 		return 10;
 	}
 
 	@Override
-	public int Max() {
+	public float Max() {
 		return 25;
 	}
