+++ src/main/java/com/robertx22/database/stat_mods/multi/offence/PhysicalDamageMulti.java
@@ -13,12 +13,12 @@ public class PhysicalDamageMulti extends StatMod {
     }
 
     @Override
-    public int Min() {
+    public float Min() {
 	return 5;
     }
 
     @Override
-    public int Max() {
+    public float Max() {
 	return 15;
     }
