+++ src/main/java/com/robertx22/database/stat_mods/percent/much_less/CrippleManaRegenPercent.java
@@ -17,12 +17,12 @@ public class CrippleManaRegenPercent extends StatMod {
     }
 
     @Override
-    public int Min() {
+    public float Min() {
 	return -25;
     }
 
     @Override
-    public int Max() {
+    public float Max() {
 	return -50;
     }
