+++ src/main/java/com/robertx22/database/stat_mods/percent/pene/BaseElePenePercent.java
@@ -5,22 +5,22 @@ import com.robertx22.uncommon.enumclasses.StatTypes;
 
 public abstract class BaseElePenePercent extends StatMod {
 
-	public BaseElePenePercent() {
-	}
+    public BaseElePenePercent() {
+    }
 
-	@Override
-	public int Min() {
-		return 10;
-	}
+    @Override
+    public float Min() {
+	return 10;
+    }
 
-	@Override
-	public int Max() {
-		return 25;
-	}
+    @Override
+    public float Max() {
+	return 25;
+    }
 
-	@Override
-	public StatTypes Type() {
-		return StatTypes.Percent;
-	}
+    @Override
+    public StatTypes Type() {
+	return StatTypes.Percent;
+    }
 
 }
