+++ src/main/java/com/robertx22/database/stat_mods/percent/spell_ele_dmg/SpellThunderDamagePercent.java
@@ -16,12 +16,12 @@ public class SpellThunderDamagePercent extends StatMod {
 	}
 
 	@Override
-	public int Min() {
+	public float Min() {
 		return 2;
 	}
 
 	@Override
-	public int Max() {
+	public float Max() {
 		return 10;
 	}
